Fix rounds type: INTEGER not REAL in SQLite

+++ apps/api/src/routes/workouts.rs
@@ -368,7 +368,7 @@ async fn embed_all_workouts(
         let name: String = row.try_get("name").map_err(AppError::from)?;
         let workout_type: Option<String> = row.try_get("workout_type").map_err(AppError::from)?;
         let duration_mins: Option<f64> = row.try_get("duration_mins").map_err(AppError::from)?;
-        let rounds: Option<f64> = row.try_get("rounds").map_err(AppError::from)?;
+        let rounds: Option<i64> = row.try_get("rounds").map_err(AppError::from)?;
         let exercises: Option<String> = row.try_get("exercise_names").map_err(AppError::from)?;
         let text = build_workout_text(&name, workout_type.as_deref(), duration_mins, rounds, exercises.as_deref());
         to_embed.push((id, text));
@@ -735,7 +735,7 @@ fn build_workout_text(
     name: &str,
     workout_type: Option<&str>,
     duration_mins: Option<f64>,
-    rounds: Option<f64>,
+    rounds: Option<i64>,
     exercises: Option<&str>,
 ) -> String {
     let mut parts = vec![format!("Workout: {name}")];
@@ -794,7 +794,7 @@ fn spawn_workout_embedding(state: &AppState, workout_id: String) {
         };
         let workout_type: Option<String> = row.try_get("workout_type").ok().flatten();
         let duration_mins: Option<f64> = row.try_get("duration_mins").ok().flatten();
-        let rounds: Option<f64> = row.try_get("rounds").ok().flatten();
+        let rounds: Option<i64> = row.try_get("rounds").ok().flatten();
         let exercises: Option<String> = row.try_get("exercise_names").ok().flatten();
 
         let text = build_workout_text(&name, workout_type.as_deref(), duration_mins, rounds, exercises.as_deref());
